requirements.txt
Flask==1.1.2
google-api-python-client==2.0.2
google-auth==1.28.0
google-cloud-pubsub==2.4.0
google-cloud-storage==1.37.0
dash==1.19.0
dash-core-components==1.15.0
dash-html-components==1.1.2
dash-renderer==1.9.0
dash-table==4.11.2
pandas==1.2.3
